+++ CHANGELOG.rst
@@ -5,6 +5,12 @@ What's New
 climpred v2.x.x (2020-06-xx)
 ============================
 
+Breaking changes
+----------------
+
+- metric :py:class:`~climpred.metrics._brier_score` now requires callable ``logical``
+  instead of ``func``. (:pr:`388`) `Aaron Spring`_.
+
 New Features
 ------------
 
@@ -50,6 +56,13 @@ Internals/Minor Fixes
   reduces the mean bias of initialized hindcasts with respect to a observation.
   See (`example <examples/decadal/bias_reduction.html>`__). (:pr:`389`) `Aaron Spring`_.
 
+
+
+Internals/Minor Fixes
+---------------------
+- Test all metrics against synthetic data. (:pr:`388`) `Aaron Spring`_.
+- metric ``threshold_brier_score`` now requires ``logical`` instead of ``func`` as
+  ``metric_kwargs``. (:pr:`388`) `Aaron Spring`_.
 
 
 climpred v2.1.0 (2020-06-08)